requirements.txt
mcp
chromadb
sentence-transformers
python-docx